requirements.txt
google-generativeai
pandas
python-dotenv
Pillow
scikit-learn
numpy